(#99) read todo/notification ids
GitHub, for some reason, gives us the ID as a string. This doesn't
make any sense, really. GitLab however gives us a number. A number in
the JavaScript sense of a number being ... a floating point
number. Whatever. We parse it into an integer and see what happens. If
we hit on any errors in the future, we may have to switch this to
strings entirely.