(#133) Add support for assigning issues on gitea
This is sort of a hack but in the same way as it was implemented for
github. Essentially the problem is that we would have to do two
independent api requests because there is no REST endpoint to add
assignees to an issue. We would first have to fetch the current
assignees, check that the assignee to add is not in the list already,
then append it and lastly submit the request to gitea. That is
insane. Not gonna happen. You'll have to live with support for only
one assignee for now.