(#140) Change leading '#' to '!' in editor input
This is necessary as almost all git forges use markdown to format
text. In markdown however, as leading '#' denotes a headline and not
comments. Comments as such do not exist in markdown. Thus, I decided
to change it to a leading '!' to allow users to put in head lines as
they wish.

If this breaks your workflow, feel free to blame me. :-)

Fixes #140

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>