(#163) Add support for NO_COLOR environment variable
This however inverts the meaning of the --no-colours flag. I change it
to --colours and -c to force colours now. This is needed because
otherwise there wouldn't be a way to force colours back despite
NO_COLOR being set in the environment.

Note: Please read the comment in config.c. I *purposefully* violate
the spec with regard to the contents of the NO_COLOR environment
variable.