Work around weird behaviour of va_arg and va_list
Apparently, as we found out the hard way, on both 32bit
FreeBSD x86 and SPARC64 Oracle Solaris you cannot simply
pass a va_list to a function and expect it to advance properly.
Reason being that va_arg must be a macro that alters the value of
vp and if you don't pass a pointer it all catches on fire and things
go horribly boom because the vp doesn't change and always returns the
first item in the va_list.