issues: Do not print Github and Gitea issues that are acutally PRs
Add a flag to gcli_issue that indicates whether a given issue is
actually a PR. In that case omit it in the output of `gcli issues` and
instead print a warning that issues have been omitted in the output.

The info message can be suppressed by supplying the -q option.