Fix: Emit an error message and print usage on missing actions
We sometimes have commands that accept a list of actions to
execute. This list gets iterated and each action including its
arguments are parsed and processed in that order. Whenever the user
gave an empty action list gcli just did nothing. We now print an error
message instead reminding the user that they need to input an
appropriate action list.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>