Move checks printing code out of library
This is a bit hackier and more involved as it looks. The problem
is the circular dependency of the definitions of the underlying
struct types. Right now we just type pun a struct with the same
layout into the right list type and dump things into it. It is the
run time's job to determine the correct routines to call on the
data.