Change forge descriptor to use new gcli_id type instead of plain int
The reasoning behind this is that if we decide to change the ID
types or if we realise that it is wrong how we handle it at the
moment we just have one typedef that affects all uses. Especially
for GitLab this seems to be important as the IDs returned by it
should be ulongs (learned from experience).

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>