(#189) Generate patches for merge requests on Gitlab
The Gitlab API does not expose an endpoint for grabbing a patch
series for all the commits of a specific version of a merge request.

Even though there is a link in the frontend rails application we
cannot use it as it will redirect us to the login page when trying
to access the patch of a private repository.

Work around this issue by manually constructing a patch that can
be used to review a MR.