feat(docs/tutorial): generate next/previous links for navigation
When generating a page, we now add next/previous links to both the header and
the footer, to help with navigating between tutorial pages.

I would have likes to stick with just using sh, however there was no clean
solution so AWK has been used to keep track of the next/previous page as well as
the page title. However the script remains POSIX compliant.