fix(docs/tutorial): Extract Full Document Titles during Page Generation
Previously, when generating HTML pages, only the first word of the page title
was used, due to the use of `awk '{print $2}'` and how we read in the toc file.
This led to incomplete titles for entries with multi-word titles.

Update the script to use a tab character as the Internal Field Separator (IFS)
whilst reading the file, ensuring full titles are accurately extracted.