(#213) Implement fetching list of pull request reviewers for Gitea and Github
This feature is somewhat incomplete due to the fact that Github
returns only the requested reviewers in the pull request data
structure.

If you attempt to request the list of reviews the returned list
contains mostly garbage that is not really related to the reviews.

For now we only print the list of requested reviewers for both of
the forges.