Implement Bugzilla OP/description of a bug report
This is a follow up for the previous change where I split out the
comments into a separate code path that skips the first comment in
the returned "comments" array.

This now adds a path that only returns the first comment in said
array. (to be specific, only the text in the first comment)