Add attachments subcommand argument parsing logic
This is very simple for now. There's only one 'get' subcommand and
not a lot of options. We will gradually expand these in the future
depending on what we need from this subcommand.

I am unsure whether adding a new subcommand is the right way to go
here but I cannot think of a better way of interfacing with attachments
right now since they are living in their own ID namespace.