Fix segmentation fault when no authheader is available in gcli_curl
This bug occured when invoking the api subcommand for bugzilla and
setting a forge type with the -t flag.

For some reason curl_slist_append calls strlen and segfaults when
the string is a null pointer.

Fix by only calling curl_slist_append when the authheader is a valid
string.