Better error messages
Whenever we print an error message to stdout we need to print the
program name too. This is because in pipes one can only distinguish
from which process an error message comes by looking at the error
messages. If it isn't clear by saying "hey I'm this program, here's
the error" one can only guess.

This is also kind of a standard on unix-y systems.