Make version announcement less noisy
Whenever an error on the command line occured and one would call
usage() it would print a ton of version and copyright noise.

Make it less noisy by adding a longversion() that prints all this
noise in this version subcommand and use a normal version() that
prints a shorter version in usage.