Remove uses of sn_sv in issues, pulls, repos, gists, forks, releases and comments
The use of these types was cumbersome and technically also incorrect.
Removing this code will allow some simplification such as a good
common set of string handling routines, a consistent API as well as
the removal of dead or rarely used code.

A common pattern we can e.g. remove is:

    escape(SV(some_c_string))

Instead we just use the jsongen set of routines.