cmd actions: add a use_pager flag and export issue and pulls actions
The use_pager flag allows us to indicate whether or not the output
should be piped through a pager such that a user can easily scroll
through it if it doesn't fit within one page of output.

The pager `less' can be configured through a command line flag (or
set in the environment variable LESS) to exit immediately if the
output fits into one page.

Exporting the action tables allows us to later reuse them in the
interactive status subsystem.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>