*: make path data an anonymous union for less verbosity
This is a minor cleanup where I made the path data an anonymous
union for better readability and less verbosity.

This hasn't been possible previously since we were using C99 and
anonymous unions are only supported starting with C11. Since we
have switched to C11 we can now use anonymous unions just fine.

Semantically this is the same. Testing with various compilers is
going to follow.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>