libgcli/github: add utilities for converting a url path to a default path
This is going to be needed to correctly resolve paths in the
interactive feature where sometimes we need to retrieve various
subitems from an object (e.g. comments of a pull request).

With Github this isn't always really easy due to the fact that they
mix semantics of a ticket ("issue") and a pull request ("request
to apply a patch stack").

Both of these have discussions but pull requests return the review
comments when asked for comments. There are more of such occurences.

Reference:     https://gitlab.com/herrhotzenplotz/gcli/-/issues/243
Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>