configure/Makefile: check for -lrt quirk on Darwin
On macOS we failed to build because the POSIX real-time extensions
are not in a seperate library "librt" but to be found in libc itself.

Since gcli links against some functions which are contained in -lrt
on some platforms we check for the host platform being a Darwin
system. If it is we simply disable -lrt linkage.

Fixes:         https://gitlab.com/herrhotzenplotz/gcli/-/issues/254
Reported-by:   botantony <https://gitlab.com/botantony>
Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>