gitlab mr: fix a double free leading to a crash on FreeBSD
I stumpled across this bug @ $WORK where I took a valgrind trace.
I then analysed this trace and also found a reproducer in the Gitlab
mirror of gcli.

Probably one of you Rust fanboys out there is now screaming and
yelling at the screen that all of this could have been so much better
if I had written this code in Rust. However then I wouldn't have
a nearly as portable and working piece of software that is maintainable
and as feature-rich and has this few dependencies.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>