configure: add a --enable-maintainer flag
The maintainer mode enables automatic rebuilding of the tutorial
manual page in case the source have changed. If this flag is set
we must be able to write the source directory. In packaging scenarios
this flag MUST NOT be set as it might cause stuff to get rebuilt
and touch the source directory.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>