Makefile.in: GNU Make sucks
It behaves differently from all other make implementations in that
it finds a dynamic build rule through the (unwritable) source
directory and attempts to regenerate the goddamn tutorial manual
page.

I thoroughly believe GNU make and its quirks is why people hate
make.

The solution for this is stupid, we hide the recipe for regenerating
the tutorial manual page template by dumping it in a rule that has
no target if in maintainer mode. Technically this is bad makefile
syntax but apparently not so who cares...

This works.

Signed-off-by: Nico Sonack <nsonack@herrhotzenplotz.de>