tests: fix integration tests for the case where no local .gcli file exists and we are not in a vcs checkout
I discovered this while working on the GoT integration in which
case I renamed the .gcli file to avoid it getting read.  In that
case the missing explicit flags in the test case definions were
causing the vcs code paths to be entered and bailing out because
it couldn't derive a valid owner/repo combination.