vcs: add support for parsing got config
This now adds full support for the game of trees vcs. We also
implement very crude matching of the host name in the apibase as a
way to figure out the forge type.

I am going to be dogfeeding this with sanitisers and full debug
info to find bugs and breakage caused by this.